requirements.txt
absl-py==1.4.0
amqp==5.2.0
asgiref==3.7.0
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.6
certifi==2023.5.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comet-ml==3.42.0
configobj==5.0.8
constantly==23.10.4
contourpy==1.1.0
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.11.0
daphne==3.0.2
dataclasses-json==0.5.14
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
dj-rest-auth==5.0.2
Django==4.2.1
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-celery-beat==2.5.0
django-channels==0.7.0
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-rest-authtoken==2.1.4
django-rest-passwordreset==1.3.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
dulwich==0.22.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
everett==3.1.0
executing==2.0.1
filelock==3.12.0
firebase-admin==6.3.0
Flask==3.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.10.0
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httplib2==0.22.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.2
imgviz==1.7.4
importlib_metadata==7.1.0
incremental==22.10.0
initialize==1.1.1
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.0
keyboard==0.13.5
kiwisolver==1.4.4
kombu==5.3.5
labelme==5.3.1
lazy_loader==0.3
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
mss==9.0.1
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysqlclient==2.1.1
natsort==8.4.0
networkx==3.1
numpy==1.25.0
oauthlib==3.2.2
onnx==1.16.1
onnxruntime==1.15.1
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==3.5.1
plyer==2.1.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.18.0
pygame==2.4.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.7.0
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.0
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySimpleGUI==4.60.5
PySocks==1.7.1
pytesseract==0.3.10
python-box==6.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-engineio==4.8.0
python-facebook-api==0.18.0
python-social-auth==0.3.6
python-socketio==5.10.0
python3-openid==3.2.0
pytweening==1.0.7
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
QtPy==2.4.0
redis==5.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.21.0
scipy==1.11.0
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==2.1.1
service-identity==24.1.0
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2023.8.30
tkinterdnd2==0.3.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
traitlets==5.14.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
twitter==1.19.6
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.2.0
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
virtualenv==20.23.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.1
win10toast==0.9
wrapt==1.15.0
wsproto==1.2.0
wurlitzer==3.1.0
zipp==3.18.1
zope.interface==6.2